Fixed perf rates not being set if perflevel is set to "default"
This commit removes the initial cache values for all rates except
the tickrate. The tickrate is used in a comparison, so we cache
that with the current server tickrate.

The bug:
To see if we need to change any rates, we check if the perf config
rates are different than the cached rates. Previously, the cache
was pre-seeded with the default level's rates. This meant that if
the perf level was set to "Default", the actual rates were never
checked and so weren't changed.